Integrate coveralls for test coverage (#81)

+++ requirements/caffe_tensorflow_install.sh
@@ -8,29 +8,31 @@ echo "Installing caffe specific dependencies"
 sudo apt-get install libprotobuf-dev libleveldb-dev libsnappy-dev libopencv-dev libhdf5-serial-dev libgflags-dev libgoogle-glog-dev liblmdb-dev protobuf-compiler
 sudo apt-get install --no-install-recommends libboost-all-dev
 
-#Download caffe
-echo "Downloading caffe"
-cd
-mkdir caffe
-cd caffe
-wget https://github.com/BVLC/caffe/archive/25391bf9e0552740af8253c6d6fd484297889a49.zip
-unzip -o 25391bf9e0552740af8253c6d6fd484297889a49.zip
-rm 25391bf9e0552740af8253c6d6fd484297889a49.zip
-mv caffe-25391bf9e0552740af8253c6d6fd484297889a49 caffe
-cd caffe
+if [ ! -d $HOME/caffe/caffe ]; then
+		#Download caffe
+		echo "Downloading caffe"
+		cd
+		mkdir caffe
+		cd caffe
+		wget https://github.com/BVLC/caffe/archive/25391bf9e0552740af8253c6d6fd484297889a49.zip
+		unzip -o 25391bf9e0552740af8253c6d6fd484297889a49.zip
+		rm 25391bf9e0552740af8253c6d6fd484297889a49.zip
+		mv caffe-25391bf9e0552740af8253c6d6fd484297889a49 caffe
+		cd caffe
 
-#Install caffe
-echo "Installing caffe"
-cp Makefile.config.example Makefile.config
-CPU_ONLY=1 USE_OPENCV=0 make all -j4
+		#Install caffe
+		echo "Installing caffe"
+		cp Makefile.config.example Makefile.config
+		CPU_ONLY=1 USE_OPENCV=0 make all -j4
 
-#Install pycaffe
-echo "Install PyCaffe"
-CPU_ONLY=1 USE_OPENCV=0 make pycaffe -j2
+		#Install pycaffe
+		echo "Install PyCaffe"
+		CPU_ONLY=1 USE_OPENCV=0 make pycaffe -j2
 
-echo "export PYTHONPATH=$(pwd)/python:$PYTHONPATH" > ~/.bash_profile
-source ~/.bash_profile
-
+		echo "export PYTHONPATH=$(pwd)/python:$PYTHONPATH" > ~/.bash_profile
+		source ~/.bash_profile
+		export PYTHONPATH=$(pwd)/python
+fi
 echo "#################### Caffe Install Complete! ####################"
 
 echo "Installing Tensorflow dependencies"
